+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -175,6 +175,11 @@ New checks
   Suggests insertion of ``std::move(...)`` to turn copy assignment operator
   calls into move assignment ones, when deemed valid and profitable.
 
+- New :doc:`readability-redundant-const
+  <clang-tidy/checks/readability/redundant-const>` check.
+
+  Detects redundant ``const`` specifiers on variable declarations.
+
 - New :doc:`readability-redundant-qualified-alias
   <clang-tidy/checks/readability/redundant-qualified-alias>` check.